user->rename in same branch fix
I have the use case where I rename in the same LDAP branch, so newParentDn is null. That's fine for ldap_rename builtin function but not for your setDn() method. Id setDn does not need to be called when renaming in the same branch, this fix should do the job.